requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pydantic
pydantic-settings
rich
requests
typing-extensions
argparse
